backend/requirements/development.txt
-r _base.txt
brunette==0.2.2
django-coverage-plugin==2.0.2
django-stubs==1.9.0
djangorestframework-stubs==1.4.0
mypy==0.931
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.0
pylint-django-stubs==0.1.0
